feat: tests reorganized (new module for base)
